# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Microservice for managing document uploads, storage, metadata, and authentication workflows\n\nFeatures:\n- Upload documents to S3 with automatic storage in DynamoDB\n- List, retrieve, and delete documents (individual or bulk)\n- Automatic file deduplication based on SHA256 hash\n- Support for multiple file types with MIME type detection\n- Document authentication workflow via RabbitMQ events\n- Transfer documents between operators (generating temporary access links for documents)\n- Event-driven architecture for user transfers and authentication results\n- Health check endpoint",
        "title": "Document Management Microservice API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/docs/documents/user/delete-all": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes all documents belonging to the authenticated user (citizen ID from JWT) and their associated files from S3 storage.\n\n## Features\n- Deletes all document metadata from DynamoDB for the authenticated user\n- Removes all physical files from S3 storage\n- Returns the count of deleted documents\n- Useful for account closure or data migration scenarios\n\n## Use Cases\n- Account closure/deletion\n- Data migration to another system\n- Bulk cleanup operations\n- GDPR/privacy compliance (right to be forgotten)\n\n## Error Codes\n- `VALIDATION_ERROR`: Invalid citizen ID\n- `PERSISTENCE_ERROR`: Failed to delete documents from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Delete all documents for the authenticated user",
                "responses": {
                    "200": {
                        "description": "All documents deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeleteAllResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error - invalid citizen ID",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeleteAllErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - database or storage error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeleteAllErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/docs/documents/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a specific document by its ID.\n\n## Features\n- Returns complete document metadata including URL for viewing/downloading\n- URL is pre-signed and ready to use in frontend viewers\n- Includes file information (size, type, hash, etc.)\n\n## Use Cases\n- Display document details in UI\n- Preview documents in viewers (PDF, images, etc.)\n- Download documents\n- Verify document integrity using hash\n\n## Error Codes\n- `NOT_FOUND`: Document with the specified ID does not exist\n- `PERSISTENCE_ERROR`: Failed to retrieve document from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/endpoints.GetResponse"
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "$ref": "#/definitions/endpoints.GetErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - database error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.GetErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a document and its associated file from S3 storage.\n\n## Features\n- Deletes document metadata from DynamoDB\n- Removes the physical file from S3 storage\n- Returns 404 if document doesn't exist\n\n## Use Cases\n- Remove unwanted documents\n- Clean up storage space\n- Comply with data deletion requests\n\n## Error Codes\n- `NOT_FOUND`: Document with the specified ID does not exist\n- `PERSISTENCE_ERROR`: Failed to delete document from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Delete a document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeleteResponse"
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeleteErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - database or storage error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.DeleteErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "endpoints.DeleteAllData": {
            "type": "object",
            "properties": {
                "deleted_count": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "endpoints.DeleteAllErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/shared.ErrorDetail"
                }
            }
        },
        "endpoints.DeleteAllResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/endpoints.DeleteAllData"
                },
                "message": {
                    "type": "string",
                    "example": "all documents deleted successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "endpoints.DeleteErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/shared.ErrorDetail"
                }
            }
        },
        "endpoints.DeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "document deleted successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "endpoints.GetErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/shared.ErrorDetail"
                }
            }
        },
        "endpoints.GetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/shared.DocumentResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "shared.DocumentResponse": {
            "type": "object",
            "properties": {
                "authentication_status": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "hash_sha256": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "shared.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request format or validation failed"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003ctoken\u003e\" (include the word Bearer followed by a space)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}